EDELIVERY-9465 new plugin - mark messages as downloaded without retrieving the payload; allow download of messages that were already downloaded and without changing the status of the message

+++ Plugin-WS/Domibus-default-ws-plugin/src/main/java/eu/domibus/plugin/ws/webservice/WebServiceImpl.java
@@ -150,7 +150,7 @@ public class WebServiceImpl implements WebServicePluginInterface {
         return response;
     }
 
-    public UserMessage getUserMessage(String messageId) {
+    public UserMessage downloadUserMessage(String messageId) {
         UserMessage userMessage;
         try {
             userMessage = wsPlugin.downloadMessage(messageId, null);
@@ -413,6 +413,38 @@ public class WebServiceImpl implements WebServicePluginInterface {
     }
 
     /**
+     * The message status is updated to downloaded (the message is not removed from the plugin table containing the pending messages so it can be downloaded using retrieveMessage
+     * @param markMessageAsDownloadedRequest
+     * @param markMessageAsDownloadedResponse
+     * @param ebMSHeaderInfo
+     * @throws eu.domibus.plugin.webService.generated.MarkMessageAsDownloadedFault
+     */
+    @Override
+    @Transactional(propagation = Propagation.REQUIRES_NEW, timeout = 300, rollbackFor = RetrieveMessageFault.class)
+    @MDCKey(value = {DomibusLogger.MDC_MESSAGE_ID, DomibusLogger.MDC_MESSAGE_ROLE, DomibusLogger.MDC_MESSAGE_ENTITY_ID}, cleanOnStart = true)
+    public void markMessageAsDownloaded(MarkMessageAsDownloadedRequest markMessageAsDownloadedRequest,
+                                        Holder<MarkMessageAsDownloadedResponse> markMessageAsDownloadedResponse,
+                                        Holder<Messaging> ebMSHeaderInfo) throws MarkMessageAsDownloadedFault {
+        boolean isMessageIdNotEmpty = StringUtils.isNotEmpty(markMessageAsDownloadedRequest.getMessageID());
+
+        if (!isMessageIdNotEmpty) {
+            LOG.error(MESSAGE_ID_EMPTY);
+            throw new MarkMessageAsDownloadedFault(MESSAGE_ID_EMPTY, webServicePluginExceptionFactory.createFault(ErrorCode.WS_PLUGIN_0007, "MessageId is empty"));
+        }
+
+        String trimmedMessageId = messageExtService.cleanMessageIdentifier(markMessageAsDownloadedRequest.getMessageID());
+        WSMessageLogEntity wsMessageLogEntity = wsMessageLogService.findByMessageId(trimmedMessageId);
+        if(wsMessageLogEntity == null) {
+            LOG.businessError(DomibusMessageCode.BUS_MSG_NOT_FOUND, trimmedMessageId);
+            throw new MarkMessageAsDownloadedFault(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]", webServicePluginExceptionFactory.createFaultMessageIdNotFound(trimmedMessageId));
+        }
+        markMessageAsDownloaded(trimmedMessageId);
+
+        markMessageAsDownloadedResponse.value =  WEBSERVICE_OF.createMarkMessageAsDownloadedResponse();
+        markMessageAsDownloadedResponse.value.setMessageID(trimmedMessageId);
+    }
+
+    /**
      * Add support for large files using DataHandler instead of byte[]
      */
     @Override
@@ -430,13 +462,14 @@ public class WebServiceImpl implements WebServicePluginInterface {
         }
 
         String trimmedMessageId = messageExtService.cleanMessageIdentifier(retrieveMessageRequest.getMessageID());
+        boolean markAsDownloaded = retrieveMessageRequest.isMarkAsDownloaded();
         WSMessageLogEntity wsMessageLogEntity = wsMessageLogService.findByMessageId(trimmedMessageId);
-        if (wsMessageLogEntity == null) {
+        if (markAsDownloaded && wsMessageLogEntity == null) {
             LOG.businessError(DomibusMessageCode.BUS_MSG_NOT_FOUND, trimmedMessageId);
             throw new RetrieveMessageFault(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]", webServicePluginExceptionFactory.createFaultMessageIdNotFound(trimmedMessageId));
         }
 
-        userMessage = getUserMessage(retrieveMessageRequest, trimmedMessageId);
+        userMessage = downloadUserMessage(trimmedMessageId, markAsDownloaded);
 
         // To avoid blocking errors during the Header's response validation
         if (StringUtils.isEmpty(userMessage.getCollaborationInfo().getAgreementRef().getValue())) {
@@ -450,34 +483,47 @@ public class WebServiceImpl implements WebServicePluginInterface {
         fillInfoPartsForLargeFiles(retrieveMessageResponse, messaging);
 
         try {
-            messageAcknowledgeExtService.acknowledgeMessageDeliveredWithUnsecureLoginAllowed(trimmedMessageId, new Timestamp(System.currentTimeMillis()), true);
+            messageAcknowledgeExtService.acknowledgeMessageDeliveredWithUnsecureLoginAllowed(trimmedMessageId, new Timestamp(System.currentTimeMillis()), markAsDownloaded);
         } catch (AuthenticationExtException | MessageAcknowledgeExtException e) {
             //if an error occurs related to the message acknowledgement do not block the download message operation
             LOG.error("Error acknowledging message [" + retrieveMessageRequest.getMessageID() + "]", e);
         }
-
-        // remove downloaded message from the plugin table containing the pending messages
-        wsMessageLogService.delete(wsMessageLogEntity);
+        if(markAsDownloaded) {
+            // remove downloaded message from the plugin table containing the pending messages
+            wsMessageLogService.delete(wsMessageLogEntity);
+        }
     }
 
-    private UserMessage getUserMessage(RetrieveMessageRequest retrieveMessageRequest, String trimmedMessageId) throws
+    private UserMessage downloadUserMessage(String trimmedMessageId, boolean markAsAcknowledged) throws
             RetrieveMessageFault {
         UserMessage userMessage;
         try {
             userMessage = wsPlugin.downloadMessage(trimmedMessageId, null);
         } catch (final MessageNotFoundException mnfEx) {
             if (LOG.isDebugEnabled()) {
-                LOG.debug(MESSAGE_NOT_FOUND_ID + retrieveMessageRequest.getMessageID() + "]", mnfEx);
+                LOG.debug(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]", mnfEx);
             }
-            LOG.error(MESSAGE_NOT_FOUND_ID + retrieveMessageRequest.getMessageID() + "]");
+            LOG.error(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]");
             throw new RetrieveMessageFault(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]", webServicePluginExceptionFactory.createFaultMessageIdNotFound(trimmedMessageId));
         }
 
         if (userMessage == null) {
-            LOG.error(MESSAGE_NOT_FOUND_ID + retrieveMessageRequest.getMessageID() + "]");
+            LOG.error(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]");
             throw new RetrieveMessageFault(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]", webServicePluginExceptionFactory.createFaultMessageIdNotFound(trimmedMessageId));
         }
         return userMessage;
+    }
+
+    private void markMessageAsDownloaded(String trimmedMessageId) throws MarkMessageAsDownloadedFault {
+        try {
+            wsPlugin.markMessageAsDownloaded(trimmedMessageId);
+        } catch (final MessageNotFoundException mnfEx) {
+            if (LOG.isDebugEnabled()) {
+                LOG.debug(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]", mnfEx);
+            }
+            LOG.error(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]");
+            throw new MarkMessageAsDownloadedFault(MESSAGE_NOT_FOUND_ID + trimmedMessageId + "]", webServicePluginExceptionFactory.createDownloadMessageFault(mnfEx));
+        }
     }
 
     private void fillInfoPartsForLargeFiles(Holder<RetrieveMessageResponse> retrieveMessageResponse, Messaging
